% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_objectives.R
\name{as.mosek_mat}
\alias{as.mosek_mat}
\title{Transforms a symmetric matrix into the form demanded by Rmosek.}
\usage{
as.mosek_mat(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A matrix in triangular sparse triplet form.
}
\description{
Transforms a symmetric matrix into the form demanded by Rmosek.
}

