% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_cross_validation.R
\name{cv.polygrams}
\alias{cv.polygrams}
\title{Performs k-fold cross-validation on the input formula.}
\usage{
cv.polygrams(formula, nfolds = 10, candidates = NULL, support = NULL,
  data = NULL, m = NULL, p = NULL, ...)
}
\arguments{
\item{formula}{The formula to work on, same semantics as for \code{polygram}.
Raw data can also be supplied.}

\item{nfolds}{Support of the data. Defaults to [0,1].}

\item{candidates}{a list of candidate polygrams. Each element of the list
must contain a vector (or integer) of splits, a vector m of Bernstein
degrees, and a vector p of derivative orders. If an integer N is supplied,
it takes the candidate list to be s = 0:N and m = 3, p = 2.}

\item{data}{an optional vector of data to be used in the formula.}

\item{m}{optional list of ms, only relevant if candidates includes only
splits.}

\item{p}{optional list of ps, only relevant if candidates includes only
splits.}

\item{...}{additional arguments passed to \code{polygram}.}
}
\value{
A \code{cv.polygram} object.
}
\description{
Performs k-fold cross-validation on the input formula.
}
\details{

}

