% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions_density.R
\name{dbernstein}
\alias{dbernstein}
\title{Evaluates a m-ary Bernstein density with specified weights.}
\usage{
dbernstein(x, lambda, log = FALSE)
}
\arguments{
\item{x}{A double where the density is evaluated.}

\item{lambda}{An m-ary vector of weights.}

\item{log}{Returns the logarithm of the density if true.}
}
\value{
A double
}
\description{
Evaluates a m-ary Bernstein density with specified weights.
}
\details{
The vector lambda gives weights to the m Bernstein
polynomials of degree m.
}

