% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions_derivative.R
\name{dbernsteinbasis}
\alias{dbernsteinbasis}
\title{Evaluates the k-th Bernstein basis density of degree m at x.}
\usage{
dbernsteinbasis(x, k, m, log = FALSE)
}
\arguments{
\item{x}{A double where the polynomial is evaluated.}

\item{k}{Specifies which among the m+1 Bernstein polynomials to use.
Runs from 0,1 ... m.}

\item{m}{The degree of the Bernstein polynomial.}

\item{log}{= FALSE Returns the logarithm of the density if true.}
}
\value{
A double
}
\description{
Evaluates the k-th Bernstein basis density of degree m at x.
}
\details{
Normalized Bernstein polynomials are positive and orthogonal on
[0,1], and integrates to 1.
}

