% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_direct_sum.R
\name{direct_sum}
\alias{direct_sum}
\title{Calculates the direct sum of matrices.}
\usage{
direct_sum(...)
}
\arguments{
\item{...}{the matrices that should be summed.}
}
\value{
The direct sum of the matrices in ... .
}
\description{
Calculates the direct sum of matrices.
}
\details{
The direct sum of matrices A, B, C, D ... is obtained by pasting
along the diagonal.
}
\examples{

A = matrix(1:9, ncol = 3)
B = matrix(1:20, ncol = 4)
Sigma = direct_sum(A, B)

}
