% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilia_loss_functions.R
\name{discrepancy}
\alias{discrepancy}
\title{Obtains the empirical discrepancy of a polygram object. If d is non-null,
it is calculated with d as the true density.}
\usage{
discrepancy(polygram_object, d = NULL)
}
\arguments{
\item{polygram_object}{A polygram object.}
}
\value{
The least squares discrepancy.
}
\description{
Obtains the empirical discrepancy of a polygram object. If d is non-null,
it is calculated with d as the true density.
}

