% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions_derivative.R
\name{dxbernstein}
\alias{dxbernstein}
\title{Evaluates the pth derivative of an m-ary Bernstein density with specified weights.}
\usage{
dxbernstein(x, lambda, p = 1, log = FALSE)
}
\arguments{
\item{x}{A double where the density is evaluated.}

\item{lambda}{An m-ary vector of weights.}

\item{p}{Desired derivative. Defaults to 1.}

\item{log}{Returns the logarithm of the density if true.}
}
\value{
A double
}
\description{
Evaluates the pth derivative of an m-ary Bernstein density with specified weights.
}
\details{
The vector lambda gives weights to the m Bernstein
polynomials of degree m.
}

