% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions_denstity.R
\name{dxbernsteinbasis}
\alias{dxbernsteinbasis}
\title{Evaluates the p-th derivative of the k-th Bernstein basis density of
degree m at x.}
\usage{
dxbernsteinbasis(x, k, m, p, log = FALSE)
}
\arguments{
\item{x}{A double where the polynomial is evaluated.}

\item{k}{Specifies which among the m+1 Bernstein polynomials to use.
Runs from 0,1 ... m.}

\item{m}{The degree of the Bernstein polynomial.}

\item{p}{Degree of derivative.}

\item{log}{= FALSE Returns the logarithm of the density if true.}
}
\value{
A double
}
\description{
Evaluates the p-th derivative of the k-th Bernstein basis density of
degree m at x.
}

