% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilia_loss_functions.R
\name{elbow_plot}
\alias{elbow_plot}
\title{Makes an elbow plot for a polygram object.}
\usage{
elbow_plot(data, lower = NULL, upper = NULL, type = "regular",
  loss = "L2", ...)
}
\arguments{
\item{data}{The underlying data}

\item{type}{Can be "regular" or "quantiles".}

\item{loss}{The type of discrepancy used. Defaults to L2, and KL will be
supported.}

\item{...}{Passed to polygram.}

\item{N}{The maximal number of bins.}
}
\value{
An elbow plot.
}
\description{
Makes an elbow plot for a polygram object.
}

