% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilia_loss_functions.R
\name{empirical_bernstein}
\alias{empirical_bernstein}
\title{Calculates the empirical or theoretical Bernstein contributions for the
data "data" (or density "d"), splits vector s, with respect to
Bernstein polynomials of order m. Used as a helper function in the
polygram_objective_vector function.}
\usage{
empirical_bernstein(data, s, m, d = NULL)
}
\arguments{
\item{data}{The data. Can be either a}

\item{s}{The vector of splits.}

\item{m}{The order of the Bernstein polynomial.}

\item{d}{Optional density, which will override the data if supplied.}
}
\value{
An (|s|-1) x (m+1)-dimensional matrix
}
\description{
Calculates the empirical or theoretical Bernstein contributions for the
data "data" (or density "d"), splits vector s, with respect to
Bernstein polynomials of order m. Used as a helper function in the
polygram_objective_vector function.
}
