% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_handle_constraints.R
\name{formulas_to_constraint_list}
\alias{formulas_to_constraint_list}
\title{A helper function for \code{polygram}. Creates a list of constraints out of
a \code{formula}, \code{ms}, \code{ps} and \code{s}.}
\usage{
formulas_to_constraint_list(form, ms, s, ps)
}
\arguments{
\item{form}{a formula}

\item{ms}{segrees of Bernstein densities}

\item{s}{split points}

\item{ps}{connectivity constraints}
}
\value{
A list containing a 1.) a constraint matrix, 2.) lower bounds and 3.)
upper bounds.
}
\description{
A helper function for \code{polygram}. Creates a list of constraints out of
a \code{formula}, \code{ms}, \code{ps} and \code{s}.
}
\details{
Currently supported options are:
\enumerate{
 \item "convex", "concave"; both together produces straight line.
 \item "decreasing", "increasing"; both together produces uniform.
 \item "symmetric"
 \item mean: Takes two "arguments", one in \code{c("=", ">=", "=<")} and a numeric
           in support.
 \item moment: Takes three "arguments", one in \code{c("=", ">=", "=<")}, a numeric
           in support, and a natural specifying the moment.
 \item quantile: Takes three "arguments", one in \code{c("=", ">=", "=<")}, a numeric
 \item quantile: Takes three "arguments", one in \code{c("=", ">=", "=<")}, a numeric
           in support, and a float in (0,1) specifying the quantile.
 }
}

