% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{get_connectivity_constraints}
\alias{get_connectivity_constraints}
\title{Returns the connectivity constraints for Cp.}
\usage{
get_connectivity_constraints(ms, s, ps, directions = NULL)
}
\arguments{
\item{s}{The vector of splits.}

\item{m}{The order of the Bernstein polynomial.}

\item{p}{Desired order of connectivity. Must be strictly less than m.}
}
\value{
A matrix of constraints.
}
\description{
Returns the connectivity constraints for Cp.
}
