% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_hard_constraints.R
\name{get_moment_constraints}
\alias{get_moment_constraints}
\title{Returns the constraints needed to enforce the pth moment to be in
the interval [a,b]}
\usage{
get_moment_constraints(m, s, moments, lower_bounds, upper_bounds)
}
\arguments{
\item{m}{The order of the Bernstein densities.}

\item{s}{The split points.}

\item{moment}{The moment desired.}

\item{a}{Lower bound for the mean}

\item{b}{Upper bound for the mean. Defaults to a.}
}
\value{
A constraint matrix.
}
\description{
Returns the constraints needed to enforce the pth moment to be in
the interval [a,b]
}

