% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_hard_constraints.R
\name{get_monotonicity_constraints}
\alias{get_monotonicity_constraints}
\title{Returns the constraints needed to enforce monotonicity of a Bernstein
polygram.}
\usage{
get_monotonicity_constraints(ms, s, monotone)
}
\arguments{
\item{s}{The split points.}

\item{monotone}{Either "decreasing" or "increasing". Can also be NULL,
but that should have been caught be now.}

\item{m}{The order of the Bernstein densities.}
}
\value{
A list containing a constraint matrix and upper and lower bounds
for the monotone
}
\description{
Returns the constraints needed to enforce monotonicity of a Bernstein
polygram.
}

