% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{get_quantile_constraints}
\alias{get_quantile_constraints}
\title{Returns the constraints needed to enforce the qth quantile to be between
the splits s_i and s_k}
\usage{
get_quantile_constraints(m, s, quantiles, lower_indices, upper_indices)
}
\arguments{
\item{m}{The order of the Bernstein densities.}

\item{s}{The split points.}

\item{quantiles}{The quantiles desired. Should be an increasing vector in (0,1).}

\item{lower_indices}{Increasing vector of lower indices for the involved split
points.}

\item{upper_indices}{Increasing vector of lower indices for the involved split
points.}
}
\value{
A list containting a constraint matrix and two vectors of lower and
upper bounds.
}
\description{
Returns the constraints needed to enforce the qth quantile to be between
the splits s_i and s_k
}
