% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_hard_constraints.R
\name{get_shape_constraints}
\alias{get_shape_constraints}
\title{Returns the constraints needed to enforce convexity / concavity of
a Bernstein histogram.}
\usage{
get_shape_constraints(ms, s, shape)
}
\arguments{
\item{s}{The split points.}

\item{shape}{Can be "convex" or "concave" or NULL.}

\item{m}{The order of the Bernstein densities.}
}
\value{
A constraint matrix.
}
\description{
Returns the constraints needed to enforce convexity / concavity of
a Bernstein histogram.
}

