% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_hard_constraints.R
\name{get_unimodal_constraints}
\alias{get_unimodal_constraints}
\title{Returns the constraints needed to enforce modality at (nu,k).}
\usage{
get_unimodal_constraints(m, s, p, unimodal, nu, k)
}
\arguments{
\item{m}{The order of the Bernstein densities.}

\item{s}{The split points.}

\item{p}{Order of connectedness.}

\item{nu}{nu-index of the mode.}

\item{k}{k-index of the mode}
}
\value{
A constraint matrix.
}
\description{
Returns the constraints needed to enforce modality at (nu,k).
}
