% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_main.R
\name{mosek2qp}
\alias{mosek2qp}
\title{Makes quadprog-compliant matrices from rmosek-compliant ones.}
\usage{
mosek2qp(constraint_matrix, lower_bounds, upper_bounds, tolerance = 10^-8)
}
\arguments{
\item{constraint_matrix}{matrix of constraints.}

\item{lower_bounds}{lower bounds for the constraints.}

\item{upper_bounds}{upper bounds for the constraints.}

\item{tolerance}{tolerance for equality checking. Defaults to 10^-8.}
}
\value{
A list containing a new constraints matrix "Amat", a vector "bvec" holding lower bounds
and a vector "mveq" holding equality constraints.
}
\description{
Makes quadprog-compliant matrices from rmosek-compliant ones.
}

