% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{objective_matrix}
\alias{objective_matrix}
\title{This function calculates the objective matrix used in the quadratic
optimization program. It is constructed recursively from its definiton.}
\usage{
objective_matrix(ms, s)
}
\arguments{
\item{ms}{an Armadillo<integer> vector containing the Bernstein degrees. Must be
of size |s| + 1}

\item{s}{an Armadillo<double> vector containing the splits, endpoints not included.}
}
\value{
an Armadillo<double> matrix of size \code{sum(ms+1)xsum(ms+1)}.
}
\description{
This function calculates the objective matrix used in the quadratic
optimization program. It is constructed recursively from its definiton.
}

