% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilia_loss_functions.R
\name{penalty}
\alias{penalty}
\title{Returns the penalty of order p for a polygram object.}
\usage{
penalty(polygram_object, p = 0)
}
\arguments{
\item{polygram_object}{A polygram object.}

\item{p}{Derivative. Only supported for p = 0 right now.}
}
\value{
The pth penalty.
}
\description{
Returns the penalty of order p for a polygram object.
}

