% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygrams.R
\name{Polygram}
\alias{Polygram}
\alias{dxpolygram}
\alias{dpolygram}
\alias{ppolygram}
\alias{rpolygram}
\title{The Polygram Class of Distributions}
\usage{
dxpolygram(x, object, support = c(0, 1), p = 1, log = FALSE)

dpolygram(x, object, support = c(0, 1), log = FALSE)

ppolygram(q, object, support = c(0, 1), lower.tail = TRUE,
  log.p = FALSE)

rpolygram(n, object)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{object}{A \code{\link{polygram}} object.}

\item{p}{vector of probabilities.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities p are given as
\code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\eqn{P[x\leX]} otherwise, \eqn{P[X<x]}.}

\item{n}{number of observations. If code{length(n) > 1}, the length is taken
to be the number required.}
}
\value{
\code{dbernstein} gives the density, \code{pbernstein} the
  distribution function, \code{qbernstein} the quantile function,
  \code{rbeta} generates random deviates, and \code{dxbernstein} gives
  the pth derivative of the density function.
}
\description{
Density, distribution function, random generation, and pth derivative of the
density for a polygram distribution object.
}
\details{
A Bernstein polygram on [a, b] is disjoint mixture of rescaled
  Bernstein densities on subintervals of [a, b]. Such polygrams can be used
  for non-parametric density estimation as an alternative to logsplines and
  kernel density estimation.
}
