% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{polygram_L2_loss}
\alias{polygram_L2_loss}
\title{Returns the true L2-loss for a weight matrix V and underlying distribution
d. Useful for theoretical investigations or simulation studies.}
\usage{
polygram_L2_loss(d, polygram_object, discrepancy = TRUE)
}
\arguments{
\item{d}{The true density.}

\item{discrepancy}{If true, only returns the discrepancy.}

\item{s}{The vector of splits. If a natural number K, s is chosen as
(1:K)/(K+1). Defaults to length(data)^(1/3)}

\item{V}{Weight matrix for the Bernstein polygram.}
}
\value{
A real number, the L2-loss.
}
\description{
Returns the true L2-loss for a weight matrix V and underlying distribution
d. Useful for theoretical investigations or simulation studies.
}
