% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_objectives.R
\name{polygram_objective_matrix}
\alias{polygram_objective_matrix}
\title{Calculates the objective matrix for use in a polygram.}
\usage{
polygram_objective_matrix(ms, s)
}
\arguments{
\item{ms}{Vector of Bernstein orders.}

\item{s}{The vector of splits.}
}
\description{
Calculates the objective matrix for use in a polygram.
}

