% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_objectives.R
\name{polygram_objective_vector}
\alias{polygram_objective_vector}
\title{Calculates the vector B used in the optimisation procedure. It supports both
data and distribution functions as input.}
\usage{
polygram_objective_vector(data, s, m, d = NULL, M = NULL)
}
\arguments{
\item{data}{The data.}

\item{s}{The vector of splits.}

\item{m}{The order of the Bernstein polynomial.}

\item{d}{An optional density.}
}
\value{
An (|s|+1)*(m+1) vector.
}
\description{
Calculates the vector B used in the optimisation procedure. It supports both
data and distribution functions as input.
}

