% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{polygram_penalty_matrix}
\alias{polygram_penalty_matrix}
\title{Calculates the penalty matrix for s, m and p.}
\usage{
polygram_penalty_matrix(s, m, p)
}
\arguments{
\item{s}{A vector of splits. Sould not include end points.}

\item{m}{The orders of the Bernstein polynomial.}

\item{p}{Derivative order.}
}
\value{
A matrix
}
\description{
Calculates the penalty matrix for s, m and p.
}
