% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilia_transformer.R
\name{polygram_transform}
\alias{polygram_transform}
\title{Changes the split vector \code{s} of a polygram with the supplied
\code{s_new}.}
\usage{
polygram_transform(polygram_object, s_new)
}
\arguments{
\item{polygram_object}{The supplied polygram_object}

\item{s_new}{The new vector of split points. Must satisfy \code{s} $\subset$
\code{s_new}.}
}
\value{
A \code{polygram} object.
}
\description{
Changes the split vector \code{s} of a polygram with the supplied
\code{s_new}.
}
\details{
Returns an equivalent polygram with the new split points. The new
polygram is equivalent to the old one in the sense that the distribution
functions are equal. It does not attempt to find inverse solutions, so
\code{s} $\subset$ \code{s_new} is required. Furthermore, this inclusion
must be exact. (No different floating points arithmetics allowed.)
}

