% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{polygram_vs_loss}
\alias{polygram_vs_loss}
\title{Calculates the L2-loss between two polygram objects. The first argument
is interpreted as the "true" polygram. Only supported when m are equal for the
moment.}
\usage{
polygram_vs_loss(polygram_object_1, polygram_object_2,
  discrepancy = FALSE)
}
\arguments{
\item{polygram_object_1}{A polygram object, interpreted as the true density.}

\item{polygram_object_2}{A polygram object.}

\item{discrepancy}{If true, only returns the discrepancy.}
}
\value{
The pth penalty.
}
\description{
Calculates the L2-loss between two polygram objects. The first argument
is interpreted as the "true" polygram. Only supported when m are equal for the
moment.
}
