% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions_random_sampling.R
\name{rpolygram}
\alias{rpolygram}
\title{Samples from a polygram.}
\usage{
rpolygram(n, polygram_object)
}
\arguments{
\item{n}{Number of desired samples.}

\item{polygram_object}{A polygram object fitted by polygram.}
}
\value{
A vector of n random samples.
}
\description{
Samples from a polygram.
}

