% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_handle_constraints.R
\name{simple_terms_to_constraint_list}
\alias{simple_terms_to_constraint_list}
\title{A helper function for \code{formulas_to_constraint_list }. Handles the simple terms.}
\usage{
simple_terms_to_constraint_list(simple_terms, ms, s, ps)
}
\arguments{
\item{ms}{segrees of Bernstein densities}

\item{s}{split points}

\item{ps}{connectivity constraints}

\item{options}{a list of options}
}
\value{
A list containing a 1.) a constraint matrix, 2.) lower bounds and 3.)
upper bounds.
}
\description{
A helper function for \code{formulas_to_constraint_list }. Handles the simple terms.
}

