% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions_denstity.R, R/base_functions_derivative.R, R/base_functions_distribution.R, R/base_functions_documentation.R, R/base_functions_quantile.R, R/base_functions_random.R
\name{dxpolygram}
\alias{base_functions}
\alias{dpolygram}
\alias{dxpolygram}
\alias{ppolygram}
\alias{qpolygram}
\alias{rpolygram}
\title{Density, distribution function, quantile function, random generation, and
pth derivative of the density for a polygram object.}
\usage{
dxpolygram(x, polygram_object, p, log = FALSE)

dpolygram(x, polygram_object, log = FALSE)

ppolygram(q, polygram_object, lower.tail = TRUE, log.p = FALSE)

qpolygram(p, polygram_object, lower.tail = TRUE, log.p = FALSE)

rpolygram(n, polygram_object)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{polygram_object}{A polygram object.}

\item{p}{vector of probabilities.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities p are given as log(p).}

\item{lower.tail}{logical; if  \code{TRUE} (default), probabilities are \eqn{P[x\leX]} otherwise, \eqn{P[X<x]}.}

\item{n}{number of observations. If code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
\code{dbernstein} gives the density, \code{pbernstein} the distribution function, \code{qbernstein}
the quantile function, \code{rbeta} generates random deviates, and \code{dxbernstein} gives the pth
derivative of the density function.
}
\description{
Density, distribution function, quantile function, random generation, and
pth derivative of the density for a polygram object.
}
\details{
A Bernstein polygram on [a, b] is disjoint mixture of rescaled Bernstein densities on subintervals
of [a, b]. Such polygrams can be used for non-parametric density estimation, as an alternative to logsplines
and kernel density estimation.
}
\section{Functions}{
\itemize{
\item \code{dxpolygram}: Calculates the pth derivative of a Bernstein polygram.

\item \code{dpolygram}: Calculates the density of a Bernstein polygram.

\item \code{ppolygram}: Calculates the distribution function of a Bernstein polygram.

\item \code{qpolygram}: Calculates quantiles of a Bernstein polygram.

\item \code{rpolygram}: Samples from a Bernstein polygram.
}}
\examples{
## Do a parametric bootstrap.
set.seed(1337)
data = rbeta(200, 2, 7)
polygram_object = polygram(data, s = 3, m = 4)
current_median = qpolygram(0.5, polygram_object)
medians = replicate(100, {
  new_data = rpolygram(200, polygram_object)
  new_median = qpolygram(0.5, polygram(new_data, s = 3, m = 4))
  new_median
})
plot(polygram(sqrt(200)*(current_median - medians), s = 4, m = 4,
              support = c(-1,1)))
}

