% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{direct_sum}
\alias{direct_sum}
\title{Calculates the direct sum of matrices.}
\usage{
direct_sum(...)
}
\arguments{
\item{...}{the matrices that should be summed.}
}
\value{
The direct sum of the matrices in \code{...}.
}
\description{
Calculates the direct sum of matrices.
}
\details{
The direct sum of matrices A, B, C, D ... is obtained by pasting
   them along the diagonal.
}
\examples{
\dontrun{
   A = matrix(1:9, ncol = 3)
   B = matrix(1:20, ncol = 4)
   Sigma = direct_sum(A, B)
}
}
\keyword{internal}
