% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_main.R
\name{polygram}
\alias{polygram}
\title{Estimates a polygram with supplied \code{data}, \code{m}, \code{p} and split points. Several
hard constraints are supported.}
\usage{
polygram(formula, s = NULL, m = NULL, p = NULL, support = NULL,
  data = NULL, method = c("quadprog", "Rmosek", "constrOptim"))
}
\arguments{
\item{formula}{The data. Can be supplied as vector of values or as a \code{formula} object,
see \code{details} for how to use the formula.}

\item{s}{The vector of split points. If a natural number \code{K}, \code{s} is chosen as
\code{(1:K)/(K+1)}. Defaults to \code{length(data)^(1/3)}. If \code{0} or \code{NULL}, a Bernstein
density is computed.}

\item{m}{The degree of the Bernstein polynomial. Can be either a number or a vector. If a number, it
is repeated |s|+1 times; if a vector, it must be of length |s|+1, where the kth element
specifies the degree of the Bernstein polynomial at the kth bin. Defaults to 3.}

\item{p}{Connectedness order.  Can be either a number or a vector. If a number, it
is repeated |s| times; if a vector, it must be of length |s|, where the kth element
specifies the degree of connectedness at the kth split. Defaults to m-1.}

\item{support}{A vector \code{c(l, u)} specifying the support of the density.
Defaults to \code{c(0, 1)} provided it contains the support of the data. Use
this option when you know something about the support of your data.}

\item{data}{an optional vector containing the data used in the formula.}

\item{method}{Optimisation method. "quadprog" (default) and "mosek" supported. Note that mosek requires
a license.}

\item{nfold}{the number of folds in cross-validation. Defaults to 5.}
}
\value{
A \code{polygram} object.
}
\description{
Estimates a polygram with supplied \code{data}, \code{m}, \code{p} and split points. Several
hard constraints are supported.
}
\details{
The \code{formula} argument takes formulas such as \code{x ~ convex + decreasing},
with the semantic that the data is on the left hand side and constraints on the right hand side.
Supported constraints are c("increasing", "decreasing", "symmetric", "concave", "convex"), but
this list will be larger soon.
}
\examples{

# A convex and decreasing density.
x = datasets::sunspot.month
plot(polygram(x ~ convex + decreasing, support = c(0, 270), s = 10))

# A crazy density.
set.seed(1984)
x = runif(11)
m = c(3,1,3,3,2,2,3,3,1,3)
p = c(1,1,2,1,0,1,2,1,1)
obj = polygram(x ~ symmetric, m = m, p = p)
plot(obj, bins = FALSE, main = "Symmetric polygram")
}

