% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polygram_objective_vector}
\alias{polygram_objective_vector}
\title{A function that calculates the objective vector of a polygram. Needed for
the quadratic optimization routine, and can be of independent interest.}
\usage{
polygram_objective_vector(data, ms, s)
}
\arguments{
\item{data}{a Armadillo<double> vector containing the data}

\item{ms}{an Armadillo<integer> vector containing the Bernstein degrees. Must be
of size |s| + 1}

\item{s_aug}{an Armadillo<double> vector containing the splits, endpoints not included.}
}
\value{
an Armadillo<double> of size \code{sum(ms+1)} that contains the
objective vector for the optimization program.
}
\description{
A function that calculates the objective vector of a polygram. Needed for
the quadratic optimization routine, and can be of independent interest.
}

